requirements.txt
ccxt==4.2.15
python-telegram-bot==20.8
numpy>=1.26.0
pandas==2.1.4
ta==0.10.2
python-dotenv==1.0.0
langchain==0.1.12
langchain-google-genai==0.0.11
httpx[socks]==0.26.0 